+++ test/orchestration/player_pool_test.dart
@@ -124,6 +124,8 @@ class _FakeBridge implements PlatformBridge {
   Future<String> deviceFingerprint() async => 'test';
   @override
   Future<int> processHeapLimitMb() async => 256;
+  @override
+  Future<void> setKeepScreenOn(bool on) async {}
 }
 
 NiumaMediaSource _src(String url) =>